fix: add resolve_api_base to sandbox-shared (DRY GHES support)
Move resolve_api_base into sandbox-shared so daytona-infra imports
it cleanly. No more local copies — all GitHub App auth lives in
sandbox-shared.
